Add status history, move file stuff to file repository
