+++ src/main/java/fi/vaylavirasto/sillari/service/FileService.java
@@ -1,17 +1,17 @@
 package fi.vaylavirasto.sillari.service;
 
 import fi.vaylavirasto.sillari.model.FileModel;
-import fi.vaylavirasto.sillari.repositories.SupervisionRepository;
+import fi.vaylavirasto.sillari.repositories.FileRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class FileService {
     @Autowired
-    SupervisionRepository supervisionRepository;
+    FileRepository fileRepository;
 
     public FileModel createFile(FileModel fileModel) {
-        Integer id = supervisionRepository.insertFileIfNotExists(fileModel);
-        return supervisionRepository.getFile(id);
+        Integer id = fileRepository.insertFileIfNotExists(fileModel);
+        return fileRepository.getFile(id);
     }
 }
